Update Q2
ive changed the available validation in the beggining so that if no boats are available the start time is changed to the smallest end time. previously the code would continue to run the originally stated time eventhough there are no available boats. secondly for when you add 30min to a time the  way the code was written it would add 30min and then go through the check endtime%100>0. therefore a number at 1000 plus 30 would be validated by the code and the end time list would be changed to 1000-30+100. so now ive changed it to starttime%10>0 so that it will only do it if the time is for example 1030.